chore(pip): Move logic to a new /python subdirectory
Create new /python subdirectories to host the pip backend and its
corresponding unit tests.

Closes https://github.com/hermetoproject/hermeto/issues/1548
