--- hermeto/core/package_managers/pip/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-only
-from hermeto.core.package_managers.pip.main import fetch_pip_source
-
-__all__ = ["fetch_pip_source"]
