+++ hermeto/core/package_managers/python/pip/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: GPL-3.0-only
+from hermeto.core.package_managers.python.pip.main import fetch_pip_source
+
+__all__ = ["fetch_pip_source"]
